Support defining per build target polyfills
